projects/3-kafka/alarms/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
requests